+++ src/app/(with-sidebar)/_components/with-sidebar-client.tsx
@@ -6,8 +6,8 @@ import { usePathname } from 'next/navigation';
 import styled from '@emotion/styled';
 import IssueHeader from '@/issues/components/header/header';
 import IssueSidebar from '@/issues/components/layout/issue-sidebar';
-import TopicHeader from '@/app/(with-sidebar)/topics/_components/header/topic-header';
-import TopicSidebar from '@/app/(with-sidebar)/topics/_components/layout/topic-sidebar';
+import TopicHeader from '@/topics/components/header/topic-header';
+import TopicSidebar from '@/topics/components/layout/topic-sidebar';
 import {
   SIDEBAR_TOGGLE_BORDER_RADIUS,
   SIDEBAR_TOGGLE_HEIGHT,
