+++ src/topics/components/issue-connection/issue-connection-line.tsx
