+++ src/topics/components/issue-node/issue-node.tsx
